Button animations
- The buttons have hover animations
- The menus close after 2 clicks outside of them
- Deleted a few obsolete scripts
