+++ Assets/Art/Animations/UI/Buttons/IdleNone.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5206d26e037954e46891e1a679866206
+guid: 1fe4732c72daf5b4e94068bb908cbbc8
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 7400000
