+++ Assets/Art/Animations/UI/Buttons/IdleRotation.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6a0c57dea3d29e04f8c7b3c437982eec
+guid: a4d04160a4c60a444931100bfb527212
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 7400000
