--- Assets/Art/Animations/UI/MouseOver.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5df15089db6144e438dc53b8c849ee1e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
