+++ Assets/Art/Sprites/UI/MenuCompra/engranaje.png.meta
@@ -133,8 +133,8 @@ TextureImporter:
         y: 10
         width: 168
         height: 173
-      alignment: 0
-      pivot: {x: 0, y: 0}
+      alignment: 9
+      pivot: {x: 0.49668354, y: 0.14251801}
       border: {x: 0, y: 0, z: 0, w: 0}
       customData: 
       outline: []
@@ -151,7 +151,7 @@ TextureImporter:
     customData: 
     physicsShape: []
     bones: []
-    spriteID: 
+    spriteID: 5b6bf9c270264ab4e8e5ea2fbf9e1e72
     internalID: 0
     vertices: []
     indices:
