+++ Assets/Resources/Prefabs/UI/Buttons.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 3c09644a35c278a44ab90ef27da3aed5
-NativeFormatImporter:
+guid: 699a76836e9d12c428019481e9b658ae
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
