+++ Assets/Resources/Prefabs/UI/UpdateWheel.prefab
@@ -309,8 +309,38 @@ PrefabInstance:
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 2948930317363838765, guid: e6a5543f2b56e1248b369cf08ac82837, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 5504038509540393685}
   m_SourcePrefab: {fileID: 100100000, guid: e6a5543f2b56e1248b369cf08ac82837, type: 3}
+--- !u!1 &1150373129403385392 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 2948930317363838765, guid: e6a5543f2b56e1248b369cf08ac82837, type: 3}
+  m_PrefabInstance: {fileID: 2817641739461310749}
+  m_PrefabAsset: {fileID: 0}
+--- !u!95 &5504038509540393685
+Animator:
+  serializedVersion: 7
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1150373129403385392}
+  m_Enabled: 1
+  m_Avatar: {fileID: 0}
+  m_Controller: {fileID: 9100000, guid: ea5fa65bbb2ba964aa4a1c0edc420db3, type: 2}
+  m_CullingMode: 0
+  m_UpdateMode: 0
+  m_ApplyRootMotion: 0
+  m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
+  m_AnimatePhysics: 0
+  m_WarningMessage: 
+  m_HasTransformHierarchy: 1
+  m_AllowConstantClipSamplingOptimization: 1
+  m_KeepAnimatorStateOnDisable: 0
+  m_WriteDefaultValuesOnDisable: 0
 --- !u!224 &4436922905172363236 stripped
 RectTransform:
   m_CorrespondingSourceObject: {fileID: 1912160551915519737, guid: e6a5543f2b56e1248b369cf08ac82837, type: 3}
@@ -415,8 +445,38 @@ PrefabInstance:
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 6201343102026380169, guid: 17961906a90a3164faca92a50eb42a9d, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 8182260591995104242}
   m_SourcePrefab: {fileID: 100100000, guid: 17961906a90a3164faca92a50eb42a9d, type: 3}
+--- !u!1 &1252729699955970734 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 6201343102026380169, guid: 17961906a90a3164faca92a50eb42a9d, type: 3}
+  m_PrefabInstance: {fileID: 5146787230692204839}
+  m_PrefabAsset: {fileID: 0}
+--- !u!95 &8182260591995104242
+Animator:
+  serializedVersion: 7
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1252729699955970734}
+  m_Enabled: 1
+  m_Avatar: {fileID: 0}
+  m_Controller: {fileID: 9100000, guid: ea5fa65bbb2ba964aa4a1c0edc420db3, type: 2}
+  m_CullingMode: 0
+  m_UpdateMode: 0
+  m_ApplyRootMotion: 0
+  m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
+  m_AnimatePhysics: 0
+  m_WarningMessage: 
+  m_HasTransformHierarchy: 1
+  m_AllowConstantClipSamplingOptimization: 1
+  m_KeepAnimatorStateOnDisable: 0
+  m_WriteDefaultValuesOnDisable: 0
 --- !u!224 &2524118734803342202 stripped
 RectTransform:
   m_CorrespondingSourceObject: {fileID: 7235726773683515997, guid: 17961906a90a3164faca92a50eb42a9d, type: 3}
