--- Assets/Src/UI/GameplayUI/BaseButton.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2a9ca60f29886db47b0947d6556a77cf
