+++ Assets/Src/UI/GameplayUI/Buttons/AButton.cs
@@ -6,15 +6,23 @@ public abstract class AButton : MonoBehaviour, IPointerEnterHandler, IPointerExi
 {
     public string name;
     public TextMeshProUGUI text;
+    private Animator _animator;
+
+    void Start()
+    {
+        _animator = GetComponent<Animator>();
+    }
     public abstract void OnPointerClick(PointerEventData eventData);
 
     public void OnPointerEnter(PointerEventData eventData)
     {
+        _animator.SetBool("hover", true);
         text.text = name;
     }
 
     public void OnPointerExit(PointerEventData eventData)
     {
+        _animator.SetBool("hover", false);
         text.text = "";
     }
 }
